modified responsiveFontSize(), added active()
responsiveFontSize() now takes at most 2 parameters, the first being the selector string and the last an optional size that overrides any 'res-n' declared sizes in an element's 'res-n' class. The function active() takes in a  selector for a <ul> element  as a parameter and decorates the active page's <a> element.